Dev coordinates

+++ DigitalDarkroom/organise_images.py
@@ -1,8 +1,6 @@
 """
 First draft:
 Extract metadata for images
-Add locations and coordinates to all images from an event or to a single image 
- - open issues: still need to differentiate between adding images to existing event vs. creating new one
 Change information fo an image or event: name or location -> add option to change the description
 Delete events or images (not tested correctly yet, but should delete images and events from database as well as from the DD folder)
 """
@@ -16,7 +14,6 @@ from geopy.geocoders import Nominatim
 import pandas as pd
 import numpy as np
 from PIL import (Image, UnidentifiedImageError)
-
 
 #######################################################################
 #Extract metadata
@@ -33,7 +30,22 @@ def get_date_from_string(date_str):
     date = np.array(date.replace(':', '-'), dtype = np.datetime64)
     return date
 
-def extract_metadata_upload(full_file_name, dest, filename):
+def get_coords(location_name):
+    """Function to get coordinates of a location 
+    input: name of a country, city, village or an address
+    output: coordinates and location name given by user, tuple
+    """
+    try:
+        geolocator = Nominatim(user_agent="DigitalDarkroom")
+        location = geolocator.geocode(location_name)
+        #print(location.address)
+        loc = (location.latitude, location.longitude, location_name)
+        return loc
+    except:
+        print("Something went wrong")
+
+
+def extract_metadata_upload(full_file_name, dest, filename, add_geo = False, single = False, coords = None):
     """ Extract metadata when images are uploaded and include it into the database image_DB
     """
     try:
@@ -63,28 +75,23 @@ def extract_metadata_upload(full_file_name, dest, filename):
                    'Date_Time':date_time,
                    'Date': date}
         config.DB.loc[filename] = pd.Series(new_row)
-        config.DB.to_pickle(os.path.join(os.environ["PROGRAM_PATH"], "image_DB.pkl")) # can be removed later -> testing
+        if add_geo:
+            if single:
+                print(f'Image name: {filename}')
+                new_location = input("Enter the location name you want to add to this image: ")
+                coords = get_coords(new_location)
+                config.DB.loc[filename, ["Latitude", "Longitude", "Location"]] = coords
+            if coords:
+                config.DB.loc[filename, ["Latitude", "Longitude", "Location"]] = coords
+        config.DB.to_pickle(os.path.join(os.environ["PROGRAM_PATH"], "image_DB.pkl")) 
         shutil.copy(full_file_name, dest)
     except UnidentifiedImageError:
         print("Not an image")
 
-########################################
-# Define functions to extract coordinates
-########################################
-def get_coords(location_name):
-    """Function to get coordinates of a location 
-    input: name of a country, city, village or an address
-    output: coordinates and location name given by user, tuple
-    """
-    try:
-        geolocator = Nominatim(user_agent="DigitalDarkroom")
-        location = geolocator.geocode(location_name)
-        #print(location.address)
-        loc = (location.latitude, location.longitude, location_name)
-        return loc
-    except:
-        print("Something went wrong")
 
+###############################################
+#### Change information from events or images
+###############################################
 
 def location_event_db(event_name, location_name):
     """Function to add or change coordinates for all pictures from the same event
@@ -101,7 +108,7 @@ def location_event_db(event_name, location_name):
 
 
 def location_image_db(image_name, location_name): 
-    """Function to add or change coordinates for all pictures from the same event
+    """Function to change coordinates for all pictures from the same event
     input: 
         image_name:     image for which coordinates will be added or changed
         location_name:  Location for which coordinates will be looked up and added to database
@@ -113,63 +120,19 @@ def location_image_db(image_name, location_name):
     config.DB.to_pickle(os.path.join(os.environ["PROGRAM_PATH"], "image_DB.pkl"))
     return config.DB
 
-
-### add geo information to events or images
-
-def add_geo_existing_event(event_name):
-    """Function to get ask user if coordinates should be added and to gather the location name
-    """
-    answer = input("Would you like to add a location to the pictures from this event? Y or N ")
-    if answer.lower() in ["yes", "y"]:
-        location_name = input("Write the name of the location that should be added - "
-                              "A country, city, village or address will do\n")
-        for image in image_list:
-            config.DB.loc[image, ["Latitude", "Longitude", "Location"]] = get_coords(location_name)
-        config.DB.to_pickle(os.path.join(os.environ["PROGRAM_PATH"], "image_DB.pkl"))
-        print(config.DB)
-        print("Location has been added to the event")
-    else:
-        print("That is ok too.")
-
-def add_geo_event(event_name):
-    """Function to get ask user if coordinates should be added and to gather the location name
-    """
-    answer = input("Would you like to add a location to the pictures from this event? Y or N ")
-    if answer.lower() in ["yes", "y"]:
-        location_name = input("Write the name of the location that should be added - "
-                              "A country, city, village or address will do\n")
-        location_event_db(event_name, location_name)
-        print(config.DB)
-        print("Location has been added to the event")
-    else:
-        print("That is ok too.")
-
-def add_geo_image(image_name):
-    """Function to get ask user if coordinates should be added and to gather the location name
-    """
-    answer = input("Would you like to add a location to the picture? Y or N ")
-    if answer.lower() in ["yes", "y"]:
-        location_name = input("Write the name of the location that should be added - "
-                              "A country, city, village or address will do\n")
-        location_image_db(image_name, location_name)
-        print(config.DB)
-        print("Location has been added to the image")
-    else:
-        print("That is ok too.")
-
-
-
-###############################################
-#### Change information from events or images
-###############################################
-
 def change_info_event():
     """Function to get ask user if coordinates should be added and to gather the location name
     """
     list_event = np.unique(config.DB["Event"].dropna())
     for event_name in list_event:
         print(event_name)
-    event = input("For which event would you like to change the information?\n")
+    event = input("For which event would you like to change the information?\n").strip()
+    while event not in list_event:
+        print("Sorry, the event was not found. Try again or type q to quit.")
+        event = input("For which event would you like to change the information?\n").strip().lower()
+        if event in ["q", "quit"]:
+            print("You decided not to change the event information.")
+            raise SystemExit
 
     change = input("What would you like to change?\n"
                    "- Name => type 'N' or name\n"
@@ -188,7 +151,7 @@ def change_info_event():
     if change.lower() in ["l", "location"]:
         location_name = input("To which country, city, village or address should the location be changed?\n")
         location_event_db(event, location_name)
-    
+
     if change.lower() in ["q", "quit"]:
         print("You decided not to change the event information.")
         raise SystemExit
@@ -201,10 +164,23 @@ def change_info_image():
     for event_name in list_event:
         print(event_name)
     event = input("In which event is the file located?\n")
+    while event not in list_event:
+        print("Sorry, the event was not found. Try again or type q to quit.")
+        event = input("In which event is the file located?\n").strip().lower()
+        if event in ["q", "quit"]:
+            print("You decided not to change the event information.")
+            raise SystemExit
     event_path = os.path.join(config.images_path, event)
-    
+    print("The following images are in the event:")
+    for file in os.listdir(event_path):
+        print(file)
+
     image_name = input("For which image would you like to change the information?") # needs to be changed -> selection available?
     file_path = os.path.join(event_path, image_name)
+    if not os.path.isfile(file_path):
+        print("Sorry, the image was not found... \n")
+        raise SystemExit
+
     change = input("What would you like to change?\n"
                    "- Name => type 'N' or name\n"
                    "- Location => type 'L' or location \n"
@@ -213,7 +189,7 @@ def change_info_image():
     if change.lower() in ["n", "name"]:
         new_name = input("What should the new name for the image be\n")
         try:
-            config.DB.loc[image_name] = new_name 
+            config.DB.rename(index={image_name:new_name}, inplace=True)
             config.DB.to_pickle(os.path.join(os.environ["PROGRAM_PATH"], "image_DB.pkl"))
             os.rename(file_path, os.path.join(event_path, new_name))
         except OSError as e:
@@ -235,14 +211,17 @@ def change_info():
                     "- Quit => type 'Q' or quit \n").lower()
     if answer in ["e", "event"]:
         change_info_event()
+        print(config.DB)
 
     if answer in ["i", "image"]:
         change_info_image()
+        print(config.DB)
 
     if answer.lower() in ["q", "quit"]:
         print("You decided not to change the event information.")
         raise SystemExit
-
+    else:
+        print("No possible option chosen")
 
 #########################################################
 # Delete event or file
@@ -251,13 +230,14 @@ def del_event():
     list_event = np.unique(config.DB["Event"].dropna())
     for event_name in list_event:
         print(event_name)
-    event = input("Which event would you like to delete?")
+    event = input("Which event would you like to delete?\n")
+    print()
     event_path = os.path.join(config.images_path, event)
-    print(event_path)
+    config.DB = config.DB.drop(config.DB[config.DB.Event == event].index)
+    print("Event deleted from database")
+    config.DB.to_pickle(os.path.join(os.environ["PROGRAM_PATH"], "image_DB.pkl"))
     try:
         shutil.rmtree(event_path)
-        config.DB = config.DB.drop(config.DB[config.DB.Event == "test"].index)
-        print(config.DB)
     except OSError as e:
         print("Error: %s - %s." % (e.filename, e.strerror))
 
@@ -267,8 +247,16 @@ def del_file():
         print(event_name)
     event = input("In which event is the file located?\n")
     file = input("Which file would you like to delete?\n")
+    print()
     event_path = os.path.join(config.images_path, event)
+    if not os.path.exists(event_path):
+        print("Sorry, the event was not found... \n")
+        raise SystemExit
+
     file_path = os.path.join(event_path, file)
+    if not os.path.isfile(file_path):
+        print("Sorry, the image was not found... \n")
+        raise SystemExit
     # Try to delete the file.
     try:
         os.remove(file_path)
@@ -285,8 +273,10 @@ def delete():
                    "- Quit => type 'Q' or 'quit'\n").lower()
     if answer in ["e", "event"]:
         del_event()
+        print(config.DB)
     if answer in ["f", "file"]:
         del_file()
+        print(config.DB)
     if answer in ['q', "quit"]:
         print("You decided not to delete anything.")
         raise SystemExit
